QString: in remove() use d->erase() for all code paths
Use the minimum of "len" and "size() - pos", logically we'll only ever
remove characters up-to the end of the array.

This eliminates one if branch, and d->erase() can handle all the use-cases.

After making this change I looked at QByteArray::remove() and it has
similar logic, so it should be correct™.

Change-Id: Ife6c552f8a3f3dda1d5a1da12b80a60790f3bae4
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>